+++ integration/BS/stiff_ode.F90
@@ -142,7 +142,11 @@ contains
     bs % eps_old = ZERO
 
     if (use_timestep_estimator) then
+#ifdef SIMPLIFIED_SDC
+       call f_bs_rhs(bs)
+#else
        call f_rhs(bs)
+#endif
        call initial_timestep(bs)
     else
        bs % dt = dt_ini
@@ -151,7 +155,11 @@ contains
     do n = 1, ode_max_steps
 
        ! Get the scaling.
+#ifdef SIMPLIFIED_SDC
+       call f_bs_rhs(bs)
+#else
        call f_rhs(bs)
+#endif
 
        if (scaling_method == 1) then
 #ifdef SIMPLIFIED_SDC
@@ -263,7 +271,12 @@ contains
 #endif
 
        ! Call the RHS, then estimate the finite difference.
+#ifdef SIMPLIFIED_SDC
+       call f_bs_rhs(bs_temp)
+#else
        call f_rhs(bs_temp)
+#endif
+
 #ifdef SIMPLIFIED_SDC
        ddydtt = (bs_temp % ydot - bs % ydot) / h
 #else
@@ -293,8 +306,15 @@ contains
   subroutine semi_implicit_extrap(bs, y, dt_tot, N_sub, y_out, ierr)
 
     !$acc routine seq
+
+#ifdef VODE
+    use linpack_module, only: dgesl, dgefa
+#else
     !$acc routine(dgesl) seq
     !$acc routine(dgefa) seq
+#endif
+
+    implicit none
 
     type (bs_t), intent(inout) :: bs
     real(rt), intent(in) :: y(bs_neqs)
@@ -331,7 +351,11 @@ contains
     enddo
 
     ! get the LU decomposition from LINPACK
+#ifdef VODE
+    call dgefa(A, ipiv, ierr_linpack)
+#else
     call dgefa(A, bs_neqs, bs_neqs, ipiv, ierr_linpack)
+#endif
     if (ierr_linpack /= 0) then
        ierr = IERR_LU_DECOMPOSITION_ERROR
     endif
@@ -353,14 +377,22 @@ contains
 #endif
 
        ! solve the first step using the LU solver
+#ifdef VODE
+       call dgesl(A, ipiv, y_out)
+#else
        call dgesl(A, bs_neqs, bs_neqs, ipiv, y_out, 0)
+#endif
 
        del(:) = y_out(:)
        bs_temp % y(:) = y(:) + del(:)
 
        t = t + h
        bs_temp % t = t
+#ifdef SIMPLIFIED_SDC
+       call f_bs_rhs(bs_temp)
+#else
        call f_rhs(bs_temp)
+#endif
 
        do n = 2, N_sub
 #ifdef SIMPLIFIED_SDC
@@ -370,14 +402,22 @@ contains
 #endif
 
           ! LU solve
+#ifdef VODE
+          call dgesl(A, ipiv, y_out)
+#else
           call dgesl(A, bs_neqs, bs_neqs, ipiv, y_out, 0)
+#endif
 
           del(:) = del(:) + TWO * y_out(:)
           bs_temp % y = bs_temp % y + del(:)
 
           t = t + h
           bs_temp % t = t
+#ifdef SIMPLIFIED_SDC
+          call f_bs_rhs(bs_temp)
+#else
           call f_rhs(bs_temp)
+#endif
        enddo
 
 #ifdef SIMPLIFIED_SDC
@@ -387,7 +427,11 @@ contains
 #endif
 
        ! last LU solve
+#ifdef VODE
+       call dgesl(A, ipiv, y_out)
+#else
        call dgesl(A, bs_neqs, bs_neqs, ipiv, y_out, 0)
+#endif
 
        ! last step
        y_out(:) = bs_temp % y(:) + y_out(:)
@@ -477,7 +521,11 @@ contains
     y_save(:) = bs % y(:)
 
     ! get the jacobian
+#ifdef SIMPLIFIED_SDC
+    call bs_jac(bs)
+#else
     call jac(bs)
+#endif
 
     if (dt /= bs % dt_next .or. bs % t /= bs % t_new) then
        bs % first = .true.
@@ -701,9 +749,13 @@ contains
     ! only of our integration variable, y
 
     !$acc routine seq
+
+#ifdef VODE
+    use linpack_module, only: dgesl, dgefa
+#else
     !$acc routine(dgesl) seq
     !$acc routine(dgefa) seq
-
+#endif
 #ifndef ACC
     use amrex_error_module, only: amrex_error
 #endif
@@ -735,7 +787,11 @@ contains
     ! note: we come in already with a RHS evalulation from the driver
 
     ! get the jacobian
+#ifdef SIMPLIFIED_SDC
+    call bs_jac(bs)
+#else
     call jac(bs)
+#endif
 
     ierr = IERR_NONE
 
@@ -758,7 +814,11 @@ contains
        enddo
        
        ! LU decomposition
+#ifdef VODE
+       call dgefa(A, ipiv, ierr_linpack)
+#else
        call dgefa(A, bs_neqs, bs_neqs, ipiv, ierr_linpack)
+#endif
        if (ierr_linpack /= 0) then
           ierr = IERR_LU_DECOMPOSITION_ERROR
        endif
@@ -771,7 +831,11 @@ contains
        g1(:) = bs % burn_s % ydot(:)
 #endif
 
+#ifdef VODE
+       call dgesl(A, ipiv, g1)
+#else
        call dgesl(A, bs_neqs, bs_neqs, ipiv, g1, 0)
+#endif
 
        ! new value of y
        bs_temp % y(:) = bs % y(:) + A21*g1(:)
@@ -779,14 +843,22 @@ contains
        
        ! get derivatives at this intermediate position and setup the next
        ! RHS
+#ifdef SIMPLIFIED_SDC
+       call f_bs_rhs(bs_temp)
+#else
        call f_rhs(bs_temp)
+#endif
 
 #ifdef SIMPLIFIED_SDC
        g2(:) = bs_temp % ydot(:) + C21*g1(:)/h
 #else
        g2(:) = bs_temp % burn_s % ydot(:) + C21*g1(:)/h
-#endif       
+#endif
+#ifdef VODE
+       call dgesl(A, ipiv, g2)
+#else
        call dgesl(A, bs_neqs, bs_neqs, ipiv, g2, 0)
+#endif
 
        ! new value of y
        bs_temp % y(:) = bs % y(:) + A31*g1(:) + A32*g2(:)
@@ -794,15 +866,22 @@ contains
 
        ! get derivatives at this intermediate position and setup the next
        ! RHS
+#ifdef SIMPLIFIED_SDC
+       call f_bs_rhs(bs_temp)
+#else
        call f_rhs(bs_temp)
+#endif
 
 #ifdef SIMPLIFIED_SDC
        g3(:) = bs_temp % ydot(:) + (C31*g1(:) + C32*g2(:))/h
 #else
        g3(:) = bs_temp % burn_s % ydot(:) + (C31*g1(:) + C32*g2(:))/h
 #endif
-       
+#ifdef VODE
+       call dgesl(A, ipiv, g3)
+#else
        call dgesl(A, bs_neqs, bs_neqs, ipiv, g3, 0)
+#endif
 
        ! our choice of parameters prevents us from needing another RHS 
        ! evaluation here
@@ -813,8 +892,11 @@ contains
 #else
        g4(:) = bs_temp % burn_s % ydot(:) + (C41*g1(:) + C42*g2(:) + C43*g3(:))/h
 #endif
-       
+#ifdef VODE
+       call dgesl(A, ipiv, g4)
+#else
        call dgesl(A, bs_neqs, bs_neqs, ipiv, g4, 0)
+#endif
 
        ! now construct our 4th order estimate of y
        bs_temp % y(:) = bs % y(:) + B1*g1(:) + B2*g2(:) + B3*g3(:) + B4*g4(:)