+++ integration/utils/network_rhs.F90
@@ -1,12 +1,14 @@
 module network_rhs_module
 
+  use amrex_fort_module, only: rt => amrex_real
+
   implicit none
 
   public
 
 contains
 
-  subroutine network_rhs(state)
+  subroutine network_rhs(state, reference_time)
 
     use actual_rhs_module, only: actual_rhs
     use burn_type_module, only: burn_t
@@ -18,19 +20,20 @@ contains
     implicit none
 
     type(burn_t), intent(inout) :: state
+    real(rt),     intent(in)    :: reference_time
 
     !$gpu
 
     call actual_rhs(state)
 
 #ifdef NONAKA_PLOT
-    call nonaka_rhs(state)
+    call nonaka_rhs(state, reference_time)
 #endif
 
   end subroutine network_rhs
 
 
-  subroutine network_jac(state)
+  subroutine network_jac(state, reference_time)
 
     use actual_rhs_module, only: actual_jac
     use burn_type_module, only: burn_t
@@ -38,6 +41,7 @@ contains
     implicit none
 
     type(burn_t), intent(inout) :: state
+    real(rt),     intent(in)    :: reference_time
 
     !$gpu